fix: Split landscape form components
